Enhancement and simplification of the smart contract language
Smart contracts will now support a more specification language set, to determine the conditions of acceptance .
This feature, enable a more consistent way to verify the conditions from the fields of the transaction either on inheritance, transaction, oracle condition.

The conditions matches are smarter, as they determine if the function calls must be applied on the condition's field or the pure function call.

Also this will lead to a better management and simplification of the parser code for the interpreter.