Switch Xandra to CQErl for performance in the DB client
With Xandra I faced some issue regarding timeout and pool management, causing GenServers to crash.

With CQErl, we can leverage asynchronous requests, so the requests will not be blocking from the connection process and
the crashes (i.e timeout) will be from the clients and not the connection handler.
Thanks to the asynchronous we can more leverage streams without blocking the connection calls.

Moreover, from benchmarks it seems more performance in term of IPS.

I also added an index in the DB to allow request by transaction address and not the chain, without duplicating the data.