Add replication attestation
To provide cross links between shards and beacon chain and to ensure
an atomic commitment of the replication is performed, validation nodes
will acknowledge confirmations from the storage nodes they are in charge
of and submit a attestation to the beacon chain and welcome node.

This notification will ensure than at least some subgroup of the entire
sharding nodes stored the transaction with an atomic commitment and
inserting their signature proofs in the beacon chain.

Clients will have to subscribe to attestation/confirmation of
replications to status as a valid transaction