Introduces genesis state and pending log
This helps to make the startup faster and ensure fault tolerance of the
memory tables information.

This is heavily inspired by some major database engine to have a commit log and a serialized
state after some threshold or time. In our case, it's the transaction of chain which trigger
serialization of inputs.